requirements.txt
setuptools
google-generativeai
google-cloud-aiplatform
google-cloud-storage
google-cloud-documentai
google-cloud-logging
requests
gcsfs
pypdf
matplotlib 
numpy
pandas
scikit-learn
fsspec
tabulate
